# Lazier unlines