Fixes #476 

Benchmark results:

before:
```
  lazy-unlines: OK (0.30s)
    547  μs ±  34 μs, 3.2 MB allocated,  23 KB copied,  38 MB peak memory
```

after:

```
  lazy-unlines: OK (0.20s)
    354  μs ±  24 μs, 2.0 MB allocated,  23 KB copied,  38 MB peak memory
```