@noughtmare 2022-01-30
Should I also change the strict version? It won't benefit from the strictness changes and a similar benchmark doesn't show very significant changes, but it does make sense to implement both functions similarly and this new implementation is simpler.