@noughtmare 2022-01-30
@sjakobi I've added a test that checks the laziness of unlines. It fails with the old implementation.