@clyring 2022-01-30
It's not directly related to the motivation for this patch, but does performance of lazy `unlines` further improve if `concat` is also rewritten as a `foldr`, so that the `List.concatMap` can potentially fuse with it? And are there any drawbacks to doing so?