@sjakobi 2022-01-30
@noughtmare What's the performance impact on the strict version?

@clyring Good question. I think it would be best to address it separately from this PR though, to keep the size of this one more manageable and to avoid a long-standing branch that might acquire merge conflicts. Feel free to record the idea on the issue tracker.