@noughtmare 2022-01-30
> What's the performance impact on the strict version?

Before:

```
    strict: OK (0.32s)
      292  μs ±  14 μs
```

After:

```
    strict: OK (0.29s)
      269  μs ±  12 μs
```